Split functionality into two get_* files
